requirements.txt
asgiref==3.5.2
boto3==1.24.32
botocore==1.27.32
Django==4.0.6
django-environ==0.9.0
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.0
importlib-metadata==4.12.0
jmespath==1.0.1
Markdown==3.3.7
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
zipp==3.8.0
